pyproject.toml
[tool.poetry]
name = "yanisbmdn-github-io"
version = "0.1.0"
description = ""
authors = ["YanisBmdn <yanis.bmdn@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flake8 = "^6.1.0"
black = "^23.10.0"


[tool.poetry.group.dev.dependencies]
pillow = "^10.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
